+++ Changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+#### 0.3.0
+- Created Basic Nodes Parser
+- Rewrote the macro to use the parser
+- Added tests
+
 #### 0.2.0
 - Refactoring expressions
 - Added Button tag